requirements.txt
beautifulsoup4
selenium
requests
pdfminer.six
